rename LintTypes.ts to lint.ts
