+++ src/providers/KangaxProvider.ts
@@ -1,4 +1,4 @@
-import { Node } from "../LintTypes";
+import { Node } from "../types";
 
 // TODO: not implemented yet.
 const KangaxProvider: Array<Node> = [];
